Move sentry_api module and update and test is_offline
