+++ src/sentry_api.rs
@@ -1,0 +1,24 @@
+use primitives::{sentry::LastApprovedResponse, ValidatorDesc};
+use reqwest::{Client, Error};
+
+pub struct SentryApi {
+    client: Client,
+}
+
+impl SentryApi {
+    pub fn new() -> Result<Self, Error> {
+        let client = Client::builder().build()?;
+
+        Ok(Self { client })
+    }
+
+    pub async fn get_last_approved(
+        &self,
+        validator: &ValidatorDesc,
+    ) -> Result<LastApprovedResponse, Error> {
+        let url = format!("{}/last-approved?withHeartbeat=true", validator.url);
+        let response = self.client.get(&url).send().await?;
+
+        response.json().await
+    }
+}
